+++ rust/backend/daemon/src/features/jni_reference_feature.rs
